Added Small Generator Module. Weapons now use energy. Pooled bullet gameobjects. Status bars for energy & hp. Disabled components no longer provide benefits to ship. Mainmodule functionality merged into module.

+++ Assets/Code/Templates/Modules/Weapons/ShotWeapons/ShotWeapon.cs
@@ -9,7 +9,7 @@ public class ShotWeapon : Weapon {
 	protected float cooldownTimer;
 
 	[Range(0, 1000)]
-	public float energyCost;
+	public int energyCost;
 
 	public void FixedUpdate(){
 		if (cooldownTimer > 0) {
